Start of version (v0.3a)
Starting to remove threads that are old and shit, with help for the legendary MrNewb! Few more PlayerPedId() changed to use the var that is set right at the top of client file instead of being independent.

I'm aware of a current issue where 9 times out of 10 you need to restart the script after server starts for it to actually work and im working on fixing that in my free time.